requirements.txt
ryu==4.34
eventlet==0.30.2
oslo.config
pyyaml